test(router): split failover execution coverage
